SC-Client supports the full contract API
